move email blueprint from yafowil.common to yafowil.email
