+++ src/yafowil/tests/test_common.py
@@ -3,7 +3,6 @@ from node.utils import UNSET
 from yafowil.base import ExtractionError
 from yafowil.base import factory
 from yafowil.common import convert_bytes
-from yafowil.compat import BYTES_TYPE
 from yafowil.compat import IS_PY2
 from yafowil.compat import UNICODE_TYPE
 from yafowil.persistence import write_mapping_writer
@@ -26,16 +25,27 @@ else:  # pragma: no cover
 class TestCommon(YafowilTestCase):
 
     def test_BC_imports(self):
+        # datatypes
         from yafowil.common import generic_emptyvalue_extractor
         from yafowil.common import generic_datatype_extractor
         from yafowil.common import DATATYPE_LABELS
-        from yafowil.common import number_extractor
 
+        # email
+        from yafowil.common import email_extractor
+
+        # field
         from yafowil.common import field_renderer
         from yafowil.common import label_renderer
         from yafowil.common import help_renderer
         from yafowil.common import error_renderer
 
+        # number
+        from yafowil.common import number_extractor
+
+        # url
+        from yafowil.common import url_extractor
+
+        # utils
         from yafowil.utils import convert_value_to_datatype
         from yafowil.utils import convert_values_to_datatype
 
@@ -3309,105 +3319,6 @@ class TestCommon(YafowilTestCase):
         model = dict()
         data.write(model)
         self.assertEqual(model, {'PWD': '1234'})
-
-    def test_email_blueprint(self):
-        # Render email input field
-        widget = factory(
-            'email',
-            name='EMAIL')
-        self.assertEqual(widget(), (
-            '<input class="email" id="input-EMAIL" name="EMAIL" '
-            'type="email" value="" />'
-        ))
-
-        # Extract not required and empty
-        data = widget.extract({'EMAIL': ''})
-        self.assertEqual(data.errors, [])
-
-        # Extract invalid email input
-        data = widget.extract({'EMAIL': 'foo@bar'})
-        self.assertEqual(
-            data.errors,
-            [ExtractionError('Input not a valid email address.')]
-        )
-
-        data = widget.extract({'EMAIL': '@bar.com'})
-        self.assertEqual(
-            data.errors,
-            [ExtractionError('Input not a valid email address.')]
-        )
-
-        # Extract valid email input
-        data = widget.extract({'EMAIL': 'foo@bar.com'})
-        self.assertEqual(data.errors, [])
-
-        # Extract required email input
-        widget = factory(
-            'email',
-            name='EMAIL',
-            props={
-                'required': 'E-Mail Address is required'
-            })
-        data = widget.extract({'EMAIL': ''})
-        self.assertEqual(
-            data.errors,
-            [ExtractionError('E-Mail Address is required')]
-        )
-
-        data = widget.extract({'EMAIL': 'foo@bar.com'})
-        self.assertEqual(data.errors, [])
-
-        # Emptyvalue
-        widget = factory(
-            'email',
-            name='EMAIL',
-            props={
-                'emptyvalue': 'foo@bar.baz',
-            })
-        data = widget.extract(request={'EMAIL': ''})
-        self.assertEqual(data.name, 'EMAIL')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, 'foo@bar.baz')
-        self.assertEqual(data.errors, [])
-
-        data = widget.extract(request={'EMAIL': 'foo@baz.bam'})
-        self.assertEqual(data.name, 'EMAIL')
-        self.assertEqual(data.value, UNSET)
-        self.assertEqual(data.extracted, 'foo@baz.bam')
-        self.assertEqual(data.errors, [])
-
-        # Datatype
-        widget = factory(
-            'email',
-            name='EMAIL',
-            props={
-                'datatype': UNICODE_TYPE
-            })
-        data = widget.extract(request={'EMAIL': u'foo@example.com'})
-        self.assertEqual(data.extracted, u'foo@example.com')
-        self.assertTrue(isinstance(data.extracted, UNICODE_TYPE))
-        self.assertFalse(isinstance(data.extracted, BYTES_TYPE))
-
-        widget = factory(
-            'email',
-            name='EMAIL',
-            props={
-                'datatype': BYTES_TYPE
-            })
-        data = widget.extract(request={'EMAIL': u'foo@example.com'})
-        self.assertEqual(data.extracted, b'foo@example.com')
-        self.assertFalse(isinstance(data.extracted, UNICODE_TYPE))
-        self.assertTrue(isinstance(data.extracted, BYTES_TYPE))
-
-        # Persist
-        widget = factory(
-            'email',
-            name='EMAIL')
-        data = widget.extract({'EMAIL': 'foo@bar.baz'})
-        model = dict()
-        data.persist_writer = write_mapping_writer
-        data.write(model)
-        self.assertEqual(model, {'EMAIL': 'foo@bar.baz'})
 
     def test_bytes_datatype_rendering_and_extraction(self):
         widget = factory(
